Create main.m
This MATLAB script generates plots of the velocity field vectors and streamlines for a given airfoil using the panelgen() function and Panel Code. The user is prompted to enter the airfoil code (a 4-digit NACA airfoil code), the freestream velocity, and the angle of attack. If the airfoil is not a NACA 2412, the script calculates and plots the lift coefficient (C_L) for the specified inputs. If the airfoil is a NACA 2412, the script reads and extracts data from an XFOIL file, and plots C_L vs AoA for the provided range of angles and number of panels (50, 100, and 200). The script includes error checking for input validity and discretises the airfoil using panelgen() and streamlineplot() functions.