Create calculator.m
This MATLAB function, 'calculator', calculates the coefficient of lift, C_L, and the strengths of the doublets, panelstrengths, for a given NACA airfoil code, number of panels, angle of attack, and free stream velocity. The function first generates and discretizes the airfoil using the 'panelgen' function. Then, it calculates the panel angle and solves the Ax = B equation using the 'cdoublet' function to find the doublet strength of each panel. The Kutta condition is fulfilled by declaring the (N+1)th row. Finally, the function solves for the doublet strength and calculates the lift coefficient. The function takes the four inputs: airfoilcode, N, AoA, and U_infinity, and produces two outputs, C_L and panelstrengths. This function can be used for aerodynamic analysis and simulation of airfoils.