Create streamlineplot.m
The streamlineplot() function is a MATLAB function that allows the user to produce any four-digit NACA airfoil with streamlines and velocity field vectors. The user specifies the four-digit airfoil code, number of panels, and angle of attack in degrees. The function returns the coefficient of lift, a streamline plot, and a velocity field vector plot. The code uses a for-loop to calculate the panel angle and to solve the Ax = B equation using the provided cdoublet.m script to find the doublet strength, panelstrengths. It then generates velocity field vectors (arrows) and streamlines around the airfoil using the meshgrid() function and identifies which U and V points are inside the airfoil and should not be plotted. The resolution of the mesh can be set higher for finer meshes, however, it will also increase the compute time.